+++ crates/llvm-context/src/polkavm/mod.rs
@@ -144,7 +144,8 @@ pub fn build_symbols(
         );
     }
 
-    let target_machine = PolkaVMTargetMachine::new(PolkaVMTarget::PVM, &OptimizerSettings::none())?;
+    let target_machine =
+        PolkaVMTargetMachine::new(PolkaVMTarget::PVM, &OptimizerSettings::none(), false)?;
     let buffer = target_machine
         .write_to_memory_buffer(&module)
         .expect("ICE: the symbols module should be valid");